Extract action from complex method
